+++ app/controllers/api/v1/tasks_controller.rb
@@ -2,26 +2,33 @@ module Api
   module V1
     class TasksController < BaseController
       def index
-        render json: project.tasks
+        tasks = policy_scope(project.tasks)
+        render json: tasks
       end
 
       def show
-        render json: project.tasks.find(params[:id])
+        task = project.tasks.find(params[:id])
+        authorize task
+        render json: task
       end
 
       def create
-        task = project.tasks.create!(task_params)
+        task = project.tasks.new(task_params)
+        authorize task
+        task.save!
         render json: task, status: :created
       end
 
       def update
         task = project.tasks.find(params[:id])
+        authorize task
         task.update!(task_params)
         render json: task
       end
 
       def destroy
         task = project.tasks.find(params[:id])
+        authorize task
         task.destroy
         head :no_content
       end
